requirements.txt
python-dotenv
pymongo
requests
beautifulsoup4
black
clickhouse-connect
reflex==0.4.9
pandas==2.2.2
numpy==1.26.4
